Add missing phrase for maternity-paternity-calculator
The text was supplied by @lutgendorff.

This adds a useful phrase when `leave_type == 'adoption'`.

Before converting to ERB, a `leave_type` of 'adoption' resulted in the
"adoption_spp_claim_link" placeholder appearing in the outcome page.

See the following URL for an example of the placeholder appearing incorrectly
(NOTE. This will no longer be visible once the switch to ERB has been deployed):
 https://www.gov.uk/maternity-paternity-calculator/y/adoption/no/2015-04-05/2015-04-06/yes/no/yes/2015-04-06/2015-01-01/2014-11-01/every_4_weeks/3000.0/usual_paydates/2015-01-01