Introduce PartYearProfitCalculator class
This is intended to capture the policy logic for the upcoming
calculate-your-part-year-profit smart answer.

The intention is that the flow will use the attribute setter methods, but as a
convenience for the tests, I've included `ActiveModel::Model` which implements
a constructor accepting a hash of attributes and values.

The aim here has been to replicate the logic of the existing
HMRC calculator [1].

[1]: http://www.hmrc.gov.uk/tools/inyearfigurecalculator/