WIP: Don't use the same calculator for every request
Note: this will probably have broken a bunch of unit tests

I was seeing a problem where `stopped_trading_on` was being set in one request
and the value was still present when making another request that didn't
explicitly set it.

I _think_ the calculator needs to be set in a `next_node_calculation` block
so that it's available to the first `next_node` block.