HACKY: Pass the state object to Date#from and Date#to
I want to be able to use the calculator object in these `from` and `to` methods
as the calculator knows the range of dates acceptable. This wasn't possible
because the `from` and `to` blocks aren't evaluated in a context that knows
about the state. I've hacked in support for passing the state so that I can get
access to the calculator object.

TODO: Work out how these date questions are actually used. Can I simplify the
called.

TODO: Do I still need to pass `state` to `parse_input` or can I get away with
just passing it to `range`? NOTE. I think that passing it to `parse_input`
allows me to rely on the `validate_input` method but I wonder how useful this
is?