(TODO titles need quotes)Use apostrophes for contractions in govspeak templates
Right single quotes and apostrophes were both used inconsistently in
contractions throughout erb templates.

After discussing this with @chrisroos we came to the conclusion that apostrophes
are better because they are easier to type on a keyboard and the end result in
HTML is the same (right single quotes) regardless of what you use in govspeak
templates.