Spike: adding govspeak mime type in Smart Answers
This allows using .govspeak as a mime type extension instead of .txt that we're
using currently. This could further lead to some work on adding custom action
view renderers that is perhaps the 'Rails Way' of handling custom view template
types.

Gotcha: I've noticed that this currently escapes `>, <, &` symbols etc.